requirements.txt
exabgp
pyYAML